Split into domain files and consolidate
This splits seccomp-dispatch.c into 6 files organized by syscall domain:
  dispatch-internal.h   shared inlines, externs, handler prototypes
  dispatch-net.c        socket/network handlers
  dispatch-id.c         UID/GID/groups identity handlers
  dispatch-exec.c       exec, mmap, mprotect, clone3
  dispatch-misc.c       time, pipe, iovec, links, ioctl, prctl

The main file retains infrastructure (path/shadow caching, guest memory
access, FD injection), core I/O/FD/directory handlers, and the dispatch
router.
